#337: create an extension with bug basing screenshot functionality which is easy to remove later
