+++ AtlasDemo/AtlasDemo/AppDelegate.swift
@@ -3,10 +3,9 @@
 //
 
 import UIKit
-import MessageUI
 
 @UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate, MFMailComposeViewControllerDelegate {
+class AppDelegate: UIResponder, UIApplicationDelegate {
 
     var window: UIWindow?
 
@@ -14,40 +13,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate, MFMailComposeViewControll
         BuddyBuildSDK.setup()
         AppSetup.configure()
 
-        NSNotificationCenter.defaultCenter().addObserverForName(UIApplicationUserDidTakeScreenshotNotification,
-                                                                object: nil,
-                                                                queue: NSOperationQueue.mainQueue()) { notification in
-            let messageController = MFMailComposeViewController()
-            messageController.setSubject("Bug Bashing")
-            messageController.mailComposeDelegate = self
-
-            if let windowRect = self.window?.bounds {
-                UIGraphicsBeginImageContext(windowRect.size)
-
-                if let ctx = UIGraphicsGetCurrentContext() {
-                    CGContextFillRect(ctx, windowRect)
-                    self.window?.layer.renderInContext(ctx)
-                    let viewimage = UIGraphicsGetImageFromCurrentImageContext()
-                    let imageData = UIImageJPEGRepresentation(viewimage, 1.0)
-
-                    if let data = imageData {
-                        messageController.addAttachmentData(data, mimeType: "image/jpeg", fileName: "Bug")
-                    }
-                }
-
-                UIGraphicsEndImageContext()
-            }
-
-            let rootVC = UIApplication.sharedApplication().keyWindow?.rootViewController
-            rootVC?.dismissViewControllerAnimated(true, completion: nil)
-            rootVC?.presentViewController(messageController, animated: true, completion: nil)
-        }
+        setupBugScreenshot()
 
         return true
-    }
-
-    func mailComposeController(controller: MFMailComposeViewController, didFinishWithResult result: MFMailComposeResult, error: NSError?) {
-        UIApplication.sharedApplication().keyWindow?.rootViewController?.dismissViewControllerAnimated(true, completion: nil)
     }
 
 }
